questionnaires: fix institute assignment, fixes #10691
git-svn-id: svn://develop.studip.de/studip/trunk@57853 c5b85792-5220-0410-ad23-ed5b6b8ad999

+++ app/controllers/questionnaire.php
@@ -518,10 +518,21 @@ class QuestionnaireController extends AuthenticatedController
                 $course_assignment->store();
             }
 
-            foreach (Request::getArray("remove_sem") as $seminar_id) {
-                if ($GLOBALS['perm']->have_studip_perm("tutor", $seminar_id)) {
+            foreach (Request::getArray('remove_sem') as $seminar_id) {
+                if ($GLOBALS['perm']->have_studip_perm('tutor', $seminar_id)) {
                     $course_assignment = QuestionnaireAssignment::findBySeminarAndQuestionnaire($seminar_id, $this->questionnaire->getId());
-                    $course_assignment->delete();
+                    if ($course_assignment) {
+                        $course_assignment->delete();
+                    }
+                }
+            }
+
+            foreach (Request::optionArray('remove_inst') as $institute_id) {
+                if ($GLOBALS['perm']->have_studip_perm('admin', $institute_id)) {
+                    $inst_assignment = QuestionnaireAssignment::findByInstituteAndQuestionnaire($institute_id, $this->questionnaire->id);
+                    if ($inst_assignment) {
+                        $inst_assignment->delete();
+                    }
                 }
             }
